// shop-with-social-distance-ui/src/app/models/product.ts
export interface Product {
  category: string;
  group: string;
  id: string;
  itemName: string;
  itemNumber: string;
  itemType: string;
  inStock: boolean;
  lowStockIndicator: boolean;
  monthlyQuotaPerUser: number;
  price: number;
  stock: number;
  toBeSoldOneItemPerOrder: boolean;
  volumePerItem: string;
  yearlyQuotaPerUser: number;
}